lab pc update
