+++ python_code/pyceres_solvers/rigid_body_solver/examples/ferret_head_example.py
@@ -258,8 +258,8 @@ def run_ferret_skull_solver(input_csv: Path, timestamps_path: Path, output_dir: 
 
 
 if __name__ == "__main__":
-    data_3d_csv = Path("/home/scholl-lab/ferret_recordings/session_2025-10-20_ferret_420_E010/full_recording/mocap_data/output_data/dlc/head_freemocap_data_by_frame.csv")
-    timestamps_npy = Path("/home/scholl-lab/ferret_recordings/session_2025-10-20_ferret_420_E010/full_recording/mocap_data/synchronized_corrected_videos/24676894_synchronized_timestamps_utc.npy")
+    data_3d_csv = Path("/home/scholl-lab/ferret_recordings/session_2025-07-11_ferret_757_EyeCamera_P43_E15__1/clips/0m_37s-1m_37s/mocap_data/output_data/dlc/head_freemocap_data_by_frame.csv")
+    timestamps_npy = Path("/home/scholl-lab/ferret_recordings/session_2025-07-11_ferret_757_EyeCamera_P43_E15__1/clips/0m_37s-1m_37s/mocap_data/synchronized_videos/24676894_synchronized_corrected_synchronized_timestamps_utc_clipped_3377_8754.npy")
     output_dir = data_3d_csv.parent.parent / "solver_output"
     output_dir.mkdir(exist_ok=True, parents=True)
     run_ferret_skull_solver(input_csv=data_3d_csv, timestamps_path=timestamps_npy, output_dir=output_dir)
